# adserver_openapi_v2.yaml
openapi: 3.0.0
info:
  description: |
    # Introduction
    Welcome to the Adserver.Online REST API reference guide.

    This API allows programmatic access to retrieve statistics, manage campaigns, users, ads, sites, zones, and more.
    It follows the [REST](https://en.wikipedia.org/wiki/REST_API) architecture and utilizes standard HTTP and JSON, making it straightforward to integrate with your systems.

    We recommend using [Postman](https://www.getpostman.com/) to explore and test the API.
    Download (button above) and import the YAML file into Postman to get started.

    You can also generate client libraries for various programming languages using [Swagger Editor](https://editor.swagger.io/).

    For SDKs and examples, visit our [GitHub repository](https://github.com/adserver-online).

    # Method groups

    In this specification, you can find 3 groups of methods: for owner, for publisher, and for advertiser.

    Using owner methods, you can access to most system components and to all users data. You must use API key created in owner panel.

    Publisher and advertiser methods provide access to role-specific components and to data for a single user only. You must use API key created in publisher or advertiser panels.

    Note: You **can't use owner's API key** to access publisher (/publish/\*) or advertiser (/advert/\*) endpoints.

...
    For example, /publish/site/123?asUser=456, where 456 is publisher's ID.

    # Rate limiting
...

...

...

    # Pagination

...

...

...

...

...

...

...

...

...
servers:
...
security:
  - Bearer: []
...
paths:
  /user:
    get:
      tags:
        - Users
      summary: List users
      operationId: getUsers
      parameters:
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: per-page
          schema:
            type: integer
        - in: query
          name: sort
          schema:
            type: string
        - $ref: '#/components/parameters/ListFilter'

      responses:
        '200':
          description: Success
          headers:
            X-Pagination-Total-Count:
              schema:
                type: integer
            X-Pagination-Page-Count:
              schema:
                type: integer
            X-Pagination-Current-Page:
              schema:
                type: integer
            X-Pagination-Per-Page:
              schema:
                type: integer
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
    post:
      tags:
        - Users
      summary: Create user
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
      responses:
        '200':
          description: Created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'

        '400':
          $ref: '#/components/responses/FormErrorResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
  /user/{id}:
    get:
      tags:
        - Users
      summary: Get user
      operationId: getUser
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
        - in: query
          name: with_balances
          schema:
            type: integer
            default: 0
            enum: [0, 1]
          description: Whether to include related balances to the response

      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'

        '404':
          description: User not found
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
    put:
      tags:
        - Users
      summary: Update user
      operationId: updateUser
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'

      responses:
        '200':
          description: Update successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'

        '400':
          $ref: '#/components/responses/FormErrorResponse'
        '404':
          description: User not found
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'

    delete:
      tags:
        - Users
      summary: Delete user
      operationId: deleteUser
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true

      responses:
        '200':
          description: Deleted successfully
        '404':
          description: User not found
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
  /stats:
    get:
      tags:
        - Reports
      summary: Statistics
      operationId: getStats
      parameters:
        - in: query
          name: dateBegin
          schema:
            type: string
          required: true
          description: Beginning of date interval
          example: '2019-01-01'

        - in: query
          name: dateEnd
          schema:
            type: string
          required: true

          description: Ending of date interval
          example: '2019-02-01'

        - in: query
          name: group
          description: Group by
          required: true
          schema:
            type: string
            enum: [day, month, year, hour, dof, dom, advertiser, campaign, price_model, campaign_category, campaign_group, ad, format, size, publisher, site, site_category, zone, zone_format, zone_size, country, country_code, postal_code, os, browser, device, brand, language, domain, domain1, referrer, ref, refid1]

        - in: query
          name: group2
          description: Extra group by. Supports the same values as "group" parameter.
          schema:
            type: string

        - in: query
          name: timezone
          description: Time zone
          schema:
            type: string
          example: UTC

        - in: query
          name: idadvertiser
          description: Filter by advertiser ID
          schema:
            type: integer

        - in: query
          name: idcampaign
          description: Filter by campaign ID
          schema:
            type: integer

        - in: query
          name: idgroup
          description: Filter by campaign group ID
          schema:
            type: integer

        - in: query
          name: idad
          description: Filter by ad ID. Supports arrays e.g. idad[]=1&idad[]=2
          schema:
            type: integer

        - in: query
          name: idpublisher
          description: Filter by publisher ID
          schema:
            type: integer

        - in: query
          name: idsite
          description: Filter by site ID
          schema:
            type: integer

        - in: query
          name: idzone
          description: Filter by zone ID. Supports arrays e.g. idzone[]=1&idzone[]=2
          schema:
            type: integer

        - in: query
          name: idcountry
          description: Filter by country geoname ID. Check https://geoname.org for IDs
          schema:
            type: integer

        - in: query
          name: iddevice
          description: Filter by device ID. Supports arrays e.g. iddevice[]=1&iddevice[]=2
          schema:
            type: integer

        - in: query
          name: idos
          description: Filter by OS ID. Supports arrays e.g. idos[]=1&idos[]=2
          schema:
            type: integer

        - in: query
          name: idbrowser
          description: Filter by browser ID. Supports arrays e.g. idbrowser[]=1&idbrowser[]=2
          schema:
            type: integer

        - in: query
          name: idadsize
          description: Filter by ad size ID. Supports arrays e.g. idadsize[]=1&idadsize[]=2
          schema:
            type: integer

        - in: query
          name: idadformat
          description: Filter by ad format ID. Supports arrays e.g. idadformat[]=1&idadformat[]=2
          schema:
            type: integer

        - in: query
          name: idzoneformat
          description: Filter by zone format ID. Supports arrays e.g. idzoneformat[]=1&idzoneformat[]=2
          schema:
            type: integer

        - in: query
          name: idcampaigncategory
          description: Filter by campaign category ID. Supports arrays e.g. idcampaigncategory[]=1&idcampaigncategory[]=2
          schema:
            type: integer

        - in: query
          name: idsitecategory
          description: Filter by site category ID. Supports arrays e.g. idsitecategory[]=1&idsitecategory[]=2
          schema:
            type: integer

        - in: query
          name: language
          description: Filter by language. Supports arrays e.g. language[]=EN&language[]=DE
          schema:
            type: string

        - in: query
          name: refid
          description: Filter by RefID
          schema:
            type: string

        - in: query
          name: with_trafq
          description: Whether to include TrafQ column to the response
          schema:
            type: integer
            enum: [0, 1]
            default: 0

        - in: query
          name: with_pub_metrics
          description: Whether to include impressions_pub and clicks_pub columns to the response
          schema:
            type: integer
            enum: [0, 1]
            default: 0

      responses:
        '200':
          description: Report created successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StatsItem'
        '400':
          $ref: '#/components/responses/FormErrorResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'

      x-code-samples:
        - lang: Curl
          source: |
            curl -X POST -H "Authorization: Bearer put_your_api_token_here" -H "Content-Type: application/json" \
              "https://api.adsrv.net/v2/stats?dateBegin=2019-01-01&dateEnd=2019-01-31&group=site"

        - lang: PHP
          source: |
            <?php
            use GuzzleHttp\Client;

            $apiToken = 'put_your_api_token_here';

            $endPoint = 'https://api.adsrv.net/v2';
            $client = new Client();

            $request = $client->get($endPoint . '/stats', [
              'headers' => [
                  'Authorization' => 'Bearer ' . $apiToken,
                  'Accept'        => 'application/json'
              ],
              'query'   => [
                  'dateBegin' => '2019-01-01',
                  'dateEnd'   => '2019-01-31',
                  'group'     => 'site'
              ]
            ]);
            $response = json_decode($request->getBody());

            var_dump($response);
  /events:
    get:
      tags:
        - Reports
      summary: Video events
      operationId: getEvents
      parameters:
        - in: query
          name: dateBegin
          schema:
            type: string
          required: true
          description: Beginning of date interval
          example: '2019-01-01'

        - in: query
          name: dateEnd
          schema:
            type: string
          required: true

          description: Ending of date interval
          example: '2019-02-01'

        - in: query
          name: report
          required: true
          schema:
            type: integer
            enum: [1, 2]
          description: >
            Report types:
             * 1 - VAST
             * 2 - Video banner

        - in: query
          name: group
          description: Group by
          required: true
          schema:
            type: string
            enum: [day, month, year, advertiser, campaign, ad, publisher, site, zone]

        - in: query
          name: timezone
          description: Time zone
          schema:
            type: string
          example: 'UTC'

        - in: query
          name: idadvertiser
          description: Filter by advertiser ID
          schema:
            type: integer

        - in: query
          name: idcampaign
          description: Filter by campaign ID
          schema:
            type: integer

        - in: query
          name: idgroup
          description: Filter by campaign group ID
          schema:
            type: integer

        - in: query
          name: idad
          description: Filter by ad ID
          schema:
            type: integer

        - in: query
          name: idpublisher
          description: Filter by publisher ID
          schema:
            type: integer

        - in: query
          name: idsite
          description: Filter by site ID
          schema:
            type: integer

        - in: query
          name: idzone
          description: Filter by zone ID
          schema:
            type: integer

      responses:
        '200':
          description: Report created successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EventsItem'
        '400':
          $ref: '#/components/responses/FormErrorResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
...
  /statement:
    get:
      tags:
        - Reports
      summary: Statement
      operationId: getStatement
      parameters:
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: per-page
          schema:
            type: integer
        - in: query
          name: sort
          schema:
            type: string
        - $ref: '#/components/parameters/ListFilter'

      responses:
        '200':
          description: Success
          headers:
            X-Pagination-Total-Count:
              schema:
                type: integer
            X-Pagination-Page-Count:
              schema:
                type: integer
            X-Pagination-Current-Page:
              schema:
                type: integer
            X-Pagination-Per-Page:
              schema:
                type: integer
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Statement'

        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
  /campaign:
    get:
      tags:
        - Campaigns
      summary: List campaigns
      operationId: getCampaigns
      parameters:
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: per-page
          schema:
            type: integer
        - in: query
          name: sort
          schema:
            type: string
        - $ref: '#/components/parameters/ListFilter'
        - in: query
          name: with_pub
          description: Whether to include assigned publishers to the response. Applies to manually assigned zones only
          schema:
            type: integer
            enum: [0, 1]
            default: 0

      responses:
        '200':
          description: Success
          headers:
            X-Pagination-Total-Count:
              schema:
                type: integer
            X-Pagination-Page-Count:
              schema:
                type: integer
            X-Pagination-Current-Page:
              schema:
                type: integer
            X-Pagination-Per-Page:
              schema:
                type: integer
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Campaign'

        '401':
          $ref: '#/components/responses/UnauthorizedResponse'

    post:
      tags:
        - Campaigns
      summary: Create campaign
      operationId: createCampaign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CampaignRequest'
      responses:
        '200':
          description: Created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Campaign'
        '400':
          $ref: '#/components/responses/FormErrorResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
  /campaign/{id}:
    get:
      tags:
        - Campaigns
      summary: Get campaign
      operationId: getCampaign
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
        - in: query
          name: with_pub
          description: Whether to include assigned publishers to the response. Applies to manually assigned zones only.
          schema:
            type: integer
            enum: [0, 1]
            default: 0

      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Campaign'

        '404':
          description: Campaign not found
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
    put:
      tags:
        - Campaigns
      summary: Update campaign
      operationId: updateCampaign
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CampaignRequest'

      responses:
        '200':
          description: Update successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Campaign'

        '400':
          $ref: '#/components/responses/FormErrorResponse'
        '404':
          description: Campaign not found
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'

      x-code-samples:
        - lang: PHP
          source: |
            <?php
            use GuzzleHttp\Client;

            $apiToken = 'put_your_api_token_here';

            $endPoint = 'https://api.adsrv.net/v2';
            $client = new Client();

            $request = $client->get($endPoint . '/campaign/123', [
              'headers' => [
                  'Authorization' => 'Bearer ' . $apiToken
              ],
              'json'   => [
                  'idrunstatus' => 4020 //pause campaign
              ]
            ]);
            $response = json_decode($request->getBody());

            var_dump($response);

    delete:
      tags:
        - Campaigns
      summary: Delete campaign
      operationId: deleteCampaign
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true

      responses:
        '200':
          description: Deleted successfully
        '404':
          description: Campaign not found
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
  /ad:
    get:
      tags:
        - Ads
      summary: List ads
      operationId: getAds
      parameters:
        - in: query
          name: idcampaign
          required: true
          schema:
            type: integer
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: per-page
          schema:
            type: integer
        - in: query
          name: sort
          schema:
            type: string
        - $ref: '#/components/parameters/ListFilter'

      responses:
        '200':
          description: Success
          headers:
            X-Pagination-Total-Count:
              schema:
                type: integer
            X-Pagination-Page-Count:
              schema:
                type: integer
            X-Pagination-Current-Page:
              schema:
                type: integer
            X-Pagination-Per-Page:
              schema:
                type: integer
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Ad'

        '401':
          $ref: '#/components/responses/UnauthorizedResponse'

    post:
      tags:
        - Ads
      summary: Create ad
      operationId: createAd
      parameters:
        - in: query
          name: idformat
          schema:
            type: integer
          required: true

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdRequest'

      responses:
        '200':
          description: Created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ad'

        '400':
          $ref: '#/components/responses/FormErrorResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
  /ad/{id}:
    get:
      tags:
        - Ads
      summary: Get ad
      operationId: getAd
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
        - in: query
          name: with_preview
          description: If 1, will return a preview html code
          schema:
            type: integer
            enum: [0, 1]
            default: 0
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ad'

        '404':
          description: Ad not found
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'

    put:
      tags:
        - Ads
      summary: Update ad
      operationId: updateAd
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdRequest'

      responses:
        '200':
          description: Update successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ad'

        '400':
          $ref: '#/components/responses/FormErrorResponse'
        '404':
          description: Ad not found
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'

    delete:
      tags:
        - Ads
      summary: Delete ad
      operationId: deleteAd
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true

      responses:
        '200':
          description: Deleted successfully
        '404':
          description: Ad not found
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
  /ad/assign:
    post:
      tags:
        - Ads
      summary: Assign ad to zones
      operationId: assignAdToZones
      parameters:
        - in: query
          name: id
          schema:
            type: integer
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - zones
              properties:
                zones:
                  type: array
                  items:
                    type: integer
      responses:
        '200':
          description: Zones assigned successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssignedZones'
        '400':
          description: Invalid request parameters
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '404':
          description: Ad not found
  /site:
    get:
      tags:
        - Sites
      summary: List sites
      operationId: getSites
      parameters:
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: per-page
          schema:
            type: integer
        - in: query
          name: sort
          schema:
            type: string
        - $ref: '#/components/parameters/ListFilter'
        - in: query
          name: with_zones
          description: Whether to include active zones list into the response
          schema:
            type: integer
            enum: [0, 1]
            default: 1
      responses:
        '200':
          description: Success
          headers:
            X-Pagination-Total-Count:
              schema:
                type: integer
            X-Pagination-Page-Count:
              schema:
                type: integer
            X-Pagination-Current-Page:
              schema:
                type: integer
            X-Pagination-Per-Page:
              schema:
                type: integer
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Site'

        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
    post:
      tags:
        - Sites
      summary: Create site
      operationId: createSite
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SiteRequest'
      responses:
        '200':
          description: Created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Site'

        '400':
          $ref: '#/components/responses/FormErrorResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
  /site/{id}:
    get:
      tags:
        - Sites
      summary: Get site
      operationId: getSite
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Site'

        '404':
          description: Site not found
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
    put:
      tags:
        - Sites
      summary: Update site
      operationId: updateSite
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SiteRequest'

      responses:
        '200':
          description: Update successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Site'

        '400':
          $ref: '#/components/responses/FormErrorResponse'
        '404':
          description: Site not found
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'

    delete:
      tags:
        - Sites
      summary: Delete site
      operationId: deleteSite
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true

      responses:
        '200':
          description: Deleted successfully
        '404':
          description: Site not found
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
...
  /zone/assign:
    post:
      tags:
        - Zones
      summary: Assign zone to ads
      operationId: assignZoneToAds
      parameters:
        - in: query
          name: id
          schema:
            type: integer
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - ads
              properties:
                ads:
                  type: array
                  items:
                    type: integer
      responses:
        '200':
          description: Ads assigned successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssignedAds'

        '400':
          description: Invalid request parameters
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '404':
          description: Zone not found
...
  /transaction/{id}:
    get:
      tags:
        - Transactions
      summary: Get transaction
      operationId: getTransaction
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'

        '404':
          description: Zone not found
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
...
components:
  securitySchemes:
    Bearer:
      type: http
      scheme: bearer
      description: |
        The most of endpoints in the API require authentication.

        Bearer authentication (also called token authentication) is an [HTTP authentication scheme](https://developer.mozilla.org/en-US/docs/Web/HTTP/Authentication) that involves security tokens called bearer tokens.
        The client must include an access token in the **Authorization** header when making requests to protected resources:
        ```
        Authorization: Bearer <token>
        ```
        You should create a token in **Account / API tokens** menu located at the top right corner of your panel.
...